// binary-type.ts
export enum BinaryType {
  Null = 'Null',
  Boolean = 'Boolean',

  UInt1 = 'UInt1',
  UInt2 = 'UInt2',
  UInt3 = 'UInt3',
  UInt4 = 'UInt4',
  UInt5 = 'UInt5',
  UInt6 = 'UInt6',
  UInt7 = 'UInt7',
  UInt8 = 'UInt8',
  UInt9 = 'UInt9',
  UInt10 = 'UInt10',
  UInt11 = 'UInt11',
  UInt12 = 'UInt12',
  UInt16 = 'UInt16',
  UInt32 = 'UInt32',

  Int2 = 'Int2',
  Int4 = 'Int4',
  Int6 = 'Int6',
  Int8 = 'Int8',
  Int10 = 'Int10',
  Int16 = 'Int16',
  Int32 = 'Int32',

  Float32 = 'Float32',
  Float64 = 'Float64',

  Rotation8 = 'Rotation8',
  ASCIIString = 'ASCIIString',
  UTF8String = 'UTF8String',
  RGB888 = 'RGB888',
  RotationFloat32 = 'RotationFloat32',
}